Use size_t consistently for array size parameters
All public-API functions now use size_t (C) / long (Java FFI) for
parameters that represent the number of elements in an array:

  pq_decoded_cosine_similarity_f32:    int  -> size_t baseOffsetsLength
  calculate_partial_sums_{dot,euclidean}_f32: int -> size_t size
    (includes internal calculate_partial_sums_f32 template and
     distance_func template helper)
  nvq_quantize_8bit:           int  -> size_t length
  nvq_loss:                    int  -> size_t length
  nvq_uniform_loss:            int  -> size_t length
  nvq_square_l2_distance_8bit: int  -> size_t length
  nvq_dot_product_8bit:        int  -> size_t length
  nvq_cosine_8bit_packed:      int  -> size_t length

NativeSimdOps.java: update FunctionDescriptor (C_INT -> C_LONG),
  method signatures (int -> long), and javadoc C-snippets (size_t).
NativeVectorUtilSupport.java: add (long) casts at all call sites
  where int values (encodedLength, size, vector.length()) are passed.

+++ jvector-native/src/main/c/jvector_simd.h
@@ -30,9 +30,9 @@ bool check_avx512_compatibility(void);
 /* PQ kernels */
 float assemble_and_sum_f32(const float* data, int dataBase, const unsigned char* baseOffsets, int baseOffsetsOffset, size_t baseOffsetsLength);
 float assemble_and_sum_pq_f32(const float* data, size_t subspaceCount, const unsigned char* baseOffsets1, int baseOffsetsOffset1, const unsigned char* baseOffsets2, int baseOffsetsOffset2, int clusterCount);
-float pq_decoded_cosine_similarity_f32(const unsigned char* baseOffsets, int baseOffsetsOffset, int baseOffsetsLength, int clusterCount, const float* partialSums, const float* aMagnitude, float bMagnitude);
-void calculate_partial_sums_euclidean_f32(const float* codebook, int codebookBase, int size, int clusterCount, const float* query, int queryOffset, float* partialSums);
-void calculate_partial_sums_dot_f32(const float* codebook, int codebookBase, int size, int clusterCount, const float* query, int queryOffset, float* partialSums);
+float pq_decoded_cosine_similarity_f32(const unsigned char* baseOffsets, int baseOffsetsOffset, size_t baseOffsetsLength, int clusterCount, const float* partialSums, const float* aMagnitude, float bMagnitude);
+void calculate_partial_sums_euclidean_f32(const float* codebook, int codebookBase, size_t size, int clusterCount, const float* query, int queryOffset, float* partialSums);
+void calculate_partial_sums_dot_f32(const float* codebook, int codebookBase, size_t size, int clusterCount, const float* query, int queryOffset, float* partialSums);
 
 /* Vector similarity kernels */
 float dot_product_f32(const float* a, size_t aoffset, const float* b, size_t boffset, size_t length);
@@ -40,12 +40,12 @@ float cosine_f32(const float* a, size_t aoffset, const float* b, size_t boffset,
 float euclidean_f32(const float* a, size_t aoffset, const float* b, size_t boffset, size_t length);
 
 /* NVQ kernels */
-void    nvq_quantize_8bit(const float* vector, int length, float alpha, float x0, float minValue, float maxValue, unsigned char* destination);
-float   nvq_loss(const float* vector, int length, float alpha, float x0, float minValue, float maxValue, int nBits);
-float   nvq_uniform_loss(const float* vector, int length, float minValue, float maxValue, int nBits);
-float   nvq_square_l2_distance_8bit(const float* vector, const unsigned char* quantized, int length, float alpha, float x0, float minValue, float maxValue);
-float   nvq_dot_product_8bit(const float* vector, const unsigned char* quantized, int length, float alpha, float x0, float minValue, float maxValue);
-int64_t nvq_cosine_8bit_packed(const float* vector, const unsigned char* quantized, int length, float alpha, float x0, float minValue, float maxValue, const float* centroid);
+void    nvq_quantize_8bit(const float* vector, size_t length, float alpha, float x0, float minValue, float maxValue, unsigned char* destination);
+float   nvq_loss(const float* vector, size_t length, float alpha, float x0, float minValue, float maxValue, int nBits);
+float   nvq_uniform_loss(const float* vector, size_t length, float minValue, float maxValue, int nBits);
+float   nvq_square_l2_distance_8bit(const float* vector, const unsigned char* quantized, size_t length, float alpha, float x0, float minValue, float maxValue);
+float   nvq_dot_product_8bit(const float* vector, const unsigned char* quantized, size_t length, float alpha, float x0, float minValue, float maxValue);
+int64_t nvq_cosine_8bit_packed(const float* vector, const unsigned char* quantized, size_t length, float alpha, float x0, float minValue, float maxValue, const float* centroid);
 
 /* Element-wise in-place vector arithmetic */
 void    add_in_place_f32(float* v1, const float* v2, size_t length);
